divided the test code into tests for graphs and layouts endpoint
