+++ tests/graphs_test.py
@@ -1,8 +1,9 @@
 from graphspace_python.graphs.classes.gsgraph import GSGraph
 from graphspace_python.api.client import GraphSpace
+from graphspace_python.api.obj.graph_response import GraphResponse
 
 
-def test_graphspace_python():
+def test_graphs_endpoint():
 	test_post_graph(name='MyTestGraph')
 	graph = test_get_graph(name='MyTestGraph')
 	test_make_graph_public(name='MyTestGraph')
@@ -11,11 +12,7 @@ def test_graphspace_python():
 	test_get_public_graphs()
 	test_get_shared_graphs()
 	test_get_my_graphs()
-	layout = test_post_graph_layout(graph_id=graph['id'])
-	test_get_my_graph_layouts(graph_id=graph['id'])
-	test_update_graph_layout(graph_id=graph['id'], layout_id=layout['id'])
-	test_get_shared_graph_layouts(graph_id=graph['id'])
-	test_delete_graph_layout(graph_id=graph['id'], layout_id=layout['id'])
+	test_get_graph_by_id()
 	test_delete_graph(name='MyTestGraph')
 
 
@@ -128,44 +125,3 @@ def test_get_my_graphs():
 	# response = graphspace.get_public_graphs(tags=['2015-bioinformatics-xtalk', 'kegg-curated-top-rank-FPs'])
 	response = graphspace.get_my_graphs()
 	assert hasattr(response, 'graphs') and len(response.graphs) > 0
-
-
-def test_get_my_graph_layouts(graph_id):
-	graphspace = GraphSpace('user1@example.com', 'user1')
-	graphspace.set_api_host('localhost:8000')
-	response = graphspace.get_my_graph_layouts(graph_id=graph_id)
-	assert hasattr(response, 'layouts') and len(response.layouts) >= 0
-
-
-def test_get_shared_graph_layouts(graph_id):
-	graphspace = GraphSpace('user1@example.com', 'user1')
-	graphspace.set_api_host('localhost:8000')
-	response = graphspace.get_shared_graph_layouts(graph_id=graph_id)
-	assert hasattr(response, 'layouts') and len(response.layouts) >= 0
-
-
-def test_post_graph_layout(graph_id):
-	graphspace = GraphSpace('user1@example.com', 'user1')
-	graphspace.set_api_host('localhost:8000')
-	response = graphspace.post_graph_layout(graph_id=graph_id, layout_name='test layout')
-	assert hasattr(response, 'layout') and response.layout.is_shared == 0
-	return response
-
-
-def test_update_graph_layout(graph_id, layout_id):
-	graphspace = GraphSpace('user1@example.com', 'user1')
-	graphspace.set_api_host('localhost:8000')
-	response = graphspace.update_graph_layout(graph_id=graph_id, layout_id=layout_id, layout_name='updated test layout', is_shared=1)
-	assert hasattr(response, 'layout') and response.layout.is_shared == 1 and response.layout.name == 'updated test layout'
-
-
-def test_delete_graph_layout(graph_id, layout_id):
-	graphspace = GraphSpace('user1@example.com', 'user1')
-	graphspace.set_api_host('localhost:8000')
-	graphspace.delete_graph_layout(graph_id=graph_id, layout_id=layout_id)
-	assert graphspace.get_graph_layout(graph_id=graph_id, layout_id=layout_id) is None
-
-# test_get_graph_by_id()
-# test_graphspace_python()
-# test_post_graph(name='MyTestGraph')
-# test_delete_graph(name='MyTestGraph')
